apps/api/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
bcrypt==4.0.1
bleach==6.2.0
click==8.3.0
dnspython==2.8.0
email-validator==2.2.0
fastapi==0.119.0
gunicorn==23.0.0
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
itsdangerous==2.2.0
orjson==3.10.12
passlib[bcrypt]==1.7.4
pydantic==2.12.0
pydantic-core==2.41.1
pydantic-settings==2.6.1
PyJWT==2.10.1
python-multipart==0.0.20
PyYAML==6.0.3
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-extensions==4.15.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
